backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.8",
    "langchain-community>=0.3.31",
    "langchain-huggingface>=0.3.1",
    "langchain-milvus>=0.1.7",
    "langchain-ollama>=0.3.10",
    "langchain-text-splitters>=0.3.11",
    "langfuse>=3.7.0",
    "langgraph>=0.6.11",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp>=1.40.0",
    "opentelemetry-instrumentation-fastapi>=0.61b0",
    "opentelemetry-instrumentation-logging>=0.61b0",
    "opentelemetry-sdk>=1.40.0",
    "prometheus-client>=0.24.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.11.0",
    "pymilvus>=2.6.11",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.39.0",
]
